factor mutable maps into a new Xml_light_utils module
There was an almost-bug in Dtd.do_prove:

   let ahash = (try find_map dtd.attribs utag with Not_found -> empty_map) in

because empty_map was defined as a global reference,

  let empty_map = ref StringMap.empty

this would reuse the same mutable empty map across calls several
proved attributes, which is the wrong behavior -- a fresh empty map
must be used each time. Happily, no mutable operation was actually
called on this map, so the bug was innocuous. It is still replaced by

    let empty_map () = ref StringMap.empty
    (*....*)
    let ahash = (try find_map dtd.attribs utag with Not_found -> empty_map ()) in